requirements.txt
Flask
gunicorn
pandas
numpy
matplotlib
seaborn
ipykernel
cassandra-driver
PYyaml
sklearn
-e .